HAR-65: send redundant Pong messages to validators
This is necessary to make sure every validators can receive
the latest list of PublicKeys.

Without this patch, the deploy.sh running locally can hardly reach
consensus for 60+ validators in one shard.

With this patch, the deploy.sh test can reach consensus on 100
validators in one shard.

Noted, we have to use the following command line to test the peer
discovery.

We have to -m <NUM_OF_VALIDATOR_IN_THE_SHARD> to stop leader from
initiating consensus if not all validators are registered. The reason is
new validator may join the shard in the middle of a consensus.

However, we have only one PublicKeys list in leader. It will be updated
once a new validator joined, and it causes inconsistency of the
PublicKeys between leader and validators and it will failure the CoSi
verification.

./deploy.sh -p -m <NUM_OF_VALIDATOR> local_config.txt

Signed-off-by: Leo Chen <leo@harmony.one>