Validate shard state in proposed blocks
For now, validators simply wait until the leader proposes something.
Going forward though, validators would have to deal with a leader
intentionally delaying resharding:

- A beacon leader should send resharding proposal as soon as DRand is
  finished.

- A regular leader should send a local resharding proposal as soon as it
  sees a master proposal from the beacon chain (received via syncing).

In both cases, there should be some sort of grace period to allow for a
short delay inserted by the leader's discretion (so that it can wait for
most nodes to catch up with DRand or beacon chain syncing).